tidied up profiling code
