+++ profiling/README.md
@@ -1,3 +1,5 @@
 # Profiling
 
 Tests we run to evaluate Piccolo performance.
+
+You need to setup a local Postgres database called 'piccolo_profile'.
