+++ requirements/profile-requirements.txt
@@ -1,2 +1,1 @@
-pyinstrument==4.1.1
 viztracer==0.15.0
