--- scripts/profile-pyinstrument.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-python profiling/setup.py
-pyinstrument profiling/profile_select.py
