--- scripts/profile-viztracer.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-python profiling/setup.py
-viztracer --log_async profiling/profile_select.py && vizviewer result.json
