+++ scripts/profile.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+python -m profiling.run_profile && vizviewer result.json
